[feature]: add ceres and g2o test
