--- doc/problems_log.md
@@ -1,5 +1,0 @@
-# Problems
-
-## std::for_each 和 std::transform 并行优缺点：
-
-
