+++ doc/todo_list.md
@@ -56,6 +56,7 @@
 - util类
   - [x] 哈希函数，实现多种哈希函数，主要是用来降低哈希碰撞的概率；
   - [x] 计算点之间的距离；
+  - [x] 实现根据输入函数曲线生成随机仿真数据的函数；
 
 - math类
   - [x] 计算一个容器中的方差，均值；
