+++ include/optimizer/slam2d.h
@@ -1,0 +1,25 @@
+#pragma once
+#include "ceres/ceres.h"
+#include "optimizer/optimizer.h"
+
+namespace Optimizer {
+
+class SLAM2D {
+ public:
+  SLAM2D();
+
+  ~SLAM2D();
+
+
+  
+
+
+  private:
+
+  
+
+
+
+};
+
+}  // namespace Optimizer
