+++ src/util/math.cc
@@ -1,3 +1,11 @@
 #include "util/math.h"
 
-namespace Utils {}  // namespace Utils
+namespace Utils {
+
+namespace Math {
+
+// 模板函数必须放到头文件中，非模板函数可以写在这个文件中
+
+}  // namespace Math
+
+}  // namespace Utils
